Other CRUD Operations including Organization, Booking, Role and User

+++ src/main/java/com/gagani/smart_booking_platform/repository/UserInfoRepository.java
@@ -11,8 +11,6 @@ public interface UserInfoRepository extends JpaRepository<UserInfo, Integer> {
 
 
     Optional<UserInfo> findUserByEmail(String email);
-
-    boolean existsByEmailAndIdNot(String email, int id);
-
+    boolean existsByEmailAndIdNot(String email, Long id);
     Optional<UserInfo> findByEmail(String username);
 }
